@@ -1,4 +1,3 @@
-from __future__ import absolute_import
 import os
 import sys
 
@@ -8,13 +7,13 @@ from camb.baseconfig import lib_import
 import argparse
 from argparse import RawTextHelpFormatter
 import ctypes
-import six
 
 
 def run_command_line():
     parser = argparse.ArgumentParser(formatter_class=RawTextHelpFormatter,
                                      description='Python command line CAMB reading parameters from a .ini file.' +
-                                                 '\n\nSample .ini files are provided in the source distribution, e.g. see inifiles/planck_2018.ini')
+                                                 '\n\nSample .ini files are provided in the source distribution, '
+                                                 'e.g. see inifiles/planck_2018.ini')
     parser.add_argument('ini_file', help='text .ini file with parameter settings')
     parser.add_argument('--validate', action='store_true',
                         help='Just validate the .ini file, don''t actually run anything')
@@ -24,7 +23,7 @@ def run_command_line():
     if not os.path.exists(args.ini_file):
         sys.exit('File not found: %s' % args.ini_file)
 
-    s = ctypes.create_string_buffer(six.b(args.ini_file))
+    s = ctypes.create_string_buffer(args.ini_file.encode("latin-1"))
 
     # Import wrapper function round fortran command line program
     CAMB_RunCommandLine = lib_import('camb', 'camb', 'CommandLineValidate' if args.validate else 'CommandLineRun')
